+++ R/aaa.R
@@ -287,8 +287,12 @@ setClass(
     iv_fDo_dgl = "integer",       # Diagonal outflow direction vector
     im_fDo     = "matrix",        # Outflow direction matrix
     im_fDi     = "matrix",        # Inflow direction matrix
+    cs_fex     = "character",     # File extension of utilised raster files
     cv_met     = "character",     # Implemented calibration quality metrics
     order      = "RPhosFateOrder" # Transport calculation order
+  ),
+  prototype = list(
+    cs_fex = ".tif"
   )
 )
 setMethod(
@@ -334,9 +338,6 @@ setMethod(
 #' @slot cv_MCl A character vector holding the names of the layers, which shall
 #'   be written to disk with the associated Monte Carlo iteration in their
 #'   filenames upon calling the appropriate methods.
-#' @slot cs_fex A character string holding the automatically determined file
-#'   extension of the provided raster files (either `".tif"` or `".img"` for
-#'   backward compatibility).
 #' @slot parameters An S4 object holding the model parameters.
 #' @slot topo An S4 object holding the raster layers related to topography in
 #'   the broader sense.
@@ -356,7 +357,6 @@ setClass(
     ls_ini     = "logical",
     is_MCi     = "integer",
     cv_MCl     = "character",
-    cs_fex     = "character",
     parameters = "RPhosFateParameters2",
     topo       = "RPhosFateTopo",
     erosion    = "RPhosFateErosion",
@@ -368,8 +368,7 @@ setClass(
     cv_dir = character(),
     ls_ini = FALSE,
     is_MCi = integer(),
-    cv_MCl = "xxt",
-    cs_fex = ".tif"
+    cv_MCl = "xxt"
   )
 )
 setMethod(
@@ -401,10 +400,6 @@ setMethod(
     }
     if (!dir.exists("Result")) {
       dir.create("Result")
-    }
-
-    if (file.exists(file.path("Input", "acc_wtd.img"))) {
-      .Object@cs_fex <- ".img"
     }
 
     .Object@substances <- new("RPhosFateSubstances", .Object)